Add context open fn and supporting mods
Fileops contains open and close wrappers from libc. fnctl is not needed
in this mod as it is already a simple function and needs no type casts
or special tricks. Also, rust does not have support for the ... argument
list outside of the ffi mod.

Errno contains errno and strerror, errno is not provded by libc. errno
fn in the Errno mod provides the error number like with C errno().
libc::__errno_location is used to access the error number.

The file descriptor of a /dev/ipath device can now  be obtained using
ipath_context_open.