Replace errno mod with io::Error and use Result<>
io::Error provides last_os_error() which is equivalent to calling errno
and strerro. One annoyance of using io::Error is that to use the
description function, you have to import error::Error which means you
have to import it under another name to get the error::Error trait.

Using io::Error reduces the need for some libc functions and allows
error numbers to be stored. io::Errors can be composed of each other,
allowing for a chain of different errors to be passed up the chain.

Will now prefer to use Result over Option when errno() will be set in
error conditions.